Great leap forward defining symbol shapes in SymbolEditControl

+++ ModelGraph/ModelGraph/Controls/Shapes/Circle.cs
@@ -10,7 +10,7 @@ namespace ModelGraph.Controls
         internal Circle()
         {
             R1 = R2 = 25;
-            DXY = new List<(sbyte dx, sbyte dy)>() { (0, 0) };
+            DXY = new List<(float dx, float dy)>() { (0, 0) };
         }
         internal Circle(int I, byte[] data) : base(I, data)
         {
